fix(Spinner): migrate the component and its documentation dependencies to TS
