+++ src/components/feedback/Spinner/interfaces/interface.Spinner.ts
@@ -1,0 +1,7 @@
+import { SizeSoinner, AppearanceSpinner } from "../types/type.Spinner";
+
+export interface ISpinnerProps {
+  size?: SizeSoinner;
+  appearance?: AppearanceSpinner;
+  isTransparent?: boolean;
+}
